+++ spec/rubocop/cops/style/parentheses_around_condition_spec.rb
@@ -6,10 +6,13 @@ module Rubocop
   module Cop
     module Style
       describe ParenthesesAroundCondition do
-        let(:pac) { ParenthesesAroundCondition.new }
+        let(:cop) { ParenthesesAroundCondition.new }
+        before do
+          ParenthesesAroundCondition.config = { 'AllowSafeAssignment' => true }
+        end
 
         it 'registers an offence for parentheses around condition' do
-          inspect_source(pac, ['if (x > 10)',
+          inspect_source(cop, ['if (x > 10)',
                                'elsif (x < 3)',
                                'end',
                                'unless (x > 10)',
@@ -23,11 +26,11 @@ module Rubocop
                                'x += 1 while (x < 10)',
                                'x += 1 until (x < 10)',
                               ])
-          expect(pac.offences.size).to eq(9)
+          expect(cop.offences.size).to eq(9)
         end
 
         it 'accepts condition without parentheses' do
-          inspect_source(pac, ['if x > 10',
+          inspect_source(cop, ['if x > 10',
                                'end',
                                'unless x > 10',
                                'end',
@@ -40,18 +43,43 @@ module Rubocop
                                'x += 1 while x < 10',
                                'x += 1 until x < 10',
                               ])
-          expect(pac.offences).to be_empty
+          expect(cop.offences).to be_empty
         end
 
         it 'is not confused by leading brace in subexpression' do
-          inspect_source(pac, ['(a > b) && other ? one : two'])
-          expect(pac.offences).to be_empty
+          inspect_source(cop, ['(a > b) && other ? one : two'])
+          expect(cop.offences).to be_empty
         end
 
         it 'is not confused by unbalanced parentheses' do
-          inspect_source(pac, ['if (a + b).c()',
+          inspect_source(cop, ['if (a + b).c()',
                                'end'])
-          expect(pac.offences).to be_empty
+          expect(cop.offences).to be_empty
+        end
+
+        context 'safe assignment is allowed' do
+          it 'accepts = in condition surrounded with braces' do
+            inspect_source(cop,
+                           ['if (test = 10)',
+                            'end'
+                           ])
+            expect(cop.offences).to be_empty
+          end
+
+        end
+
+        context 'safe assignment is not allowed' do
+          before do
+            ParenthesesAroundCondition.config['AllowSafeAssignment'] = false
+          end
+
+          it 'does not accepts = in condition surrounded with braces' do
+            inspect_source(cop,
+                           ['if (test = 10)',
+                            'end'
+                           ])
+            expect(cop.offences.size).to eq(1)
+          end
         end
       end
     end